+++ iftttie/services/base.py
@@ -1,17 +1,14 @@
 from __future__ import annotations
 
 from abc import ABCMeta, abstractmethod
-from asyncio import Queue
 from typing import Any
 
-from aiohttp import ClientSession
-
-from iftttie.types import Update
+from iftttie import web
 
 
 class BaseService(metaclass=ABCMeta):
     @abstractmethod
-    async def run(self, client_session: ClientSession, event_queue: Queue[Update], **kwargs: Any):
+    async def run(self, context: web.Context, **kwargs: Any):
         """
         Runs the service. May raise unhandled exceptions.
         """